basic fullscreen video  functionality working on ios. Android plays videos but needs controls.
